test(twitter): split text renderer coverage
